+++ src/main/java/roomescape/business/service/ReservationTimeService.java
@@ -9,9 +9,9 @@ import roomescape.business.ReservationTime;
 import roomescape.exception.ReservationTimeException;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationTimeRepository;
-import roomescape.presentation.dto.AvailableTimesResponseDto;
-import roomescape.presentation.dto.ReservationTimeRequestDto;
-import roomescape.presentation.dto.ReservationTimeResponseDto;
+import roomescape.presentation.dto.response.AvailableTimesResponseDto;
+import roomescape.presentation.dto.request.ReservationTimeRequestDto;
+import roomescape.presentation.dto.response.ReservationTimeResponseDto;
 
 @Service
 @Transactional
