+++ src/main/java/roomescape/presentation/controller/admin/page/AdminPageController.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.page;
+package roomescape.presentation.controller.admin.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
