+++ src/main/java/roomescape/presentation/controller/user/ReservationThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.presentation;
+package roomescape.presentation.controller.user;
 
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.presentation.dto.ReservationThemeRequestDto;
-import roomescape.presentation.dto.ReservationThemeResponseDto;
+import roomescape.presentation.dto.request.ReservationThemeRequestDto;
+import roomescape.presentation.dto.response.ReservationThemeResponseDto;
 import roomescape.business.service.ReservationThemeService;
 
 @RestController
