+++ src/main/java/roomescape/presentation/dto/response/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.dto;
+package roomescape.presentation.dto.response;
 
 import java.time.LocalDate;
