+++ src/main/java/roomescape/presentation/dto/response/ReservationTimeResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.dto;
+package roomescape.presentation.dto.response;
 
 import java.time.LocalTime;
