+++ src/test/java/roomescape/business/service/ReservationTimeServiceTest.java
@@ -21,9 +21,8 @@ import roomescape.exception.ReservationTimeException;
 import roomescape.persistence.MemberRepository;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationTimeRepository;
-import roomescape.presentation.dto.MemberResponseDto;
-import roomescape.presentation.dto.ReservationTimeRequestDto;
-import roomescape.presentation.dto.ReservationTimeResponseDto;
+import roomescape.presentation.dto.request.ReservationTimeRequestDto;
+import roomescape.presentation.dto.response.ReservationTimeResponseDto;
 
 class ReservationTimeServiceTest {
